+++ bot/src/main/kotlin/org/epilink/bot/discord/LinkRoleManager.kt
@@ -217,11 +217,13 @@ internal class LinkRoleManagerImpl : LinkRoleManager, KoinComponent {
         logger.debug { "Updating roles for $discordId in $guildId: they should have ${roles.joinToString(", ")}" }
         val serverConfig = config.getConfigForGuild(guildId)
         val toObtain = roles.mapNotNull { serverConfig.roles[it] }.toSet()
-        val toRemove = serverConfig.roles.values.toSet().minus(toObtain)
+        val toRemove = serverConfig.roles
+            .filterKeys { it !in serverConfig.stickyRoles && it !in config.stickyRoles}
+            .values - toObtain
         logger.debug {
             "For $discordId in $guildId: remove ${toRemove.joinToString(", ")} and add ${toObtain.joinToString(", ")}"
         }
-        facade.manageRoles(discordId, guildId, toObtain, toRemove)
+        facade.manageRoles(discordId, guildId, toObtain, toRemove.toSet())
     }
 
     @OptIn(UsesTrueIdentity::class)